add a command line parameter to force or not the update of the files

+++ 01-get_contracts.py
@@ -14,6 +14,7 @@ import json
 import os
 import re
 from utils import print_progress
+import argparse
 
 CONTRACTS_JSON_URL_EU = "https://opendata.euskadi.eus/contenidos/ds_contrataciones/contrataciones_admin_2021/opendata/kontratuak.json"
 CONTRACTS_JSON_URL_ES = "https://opendata.euskadi.eus/contenidos/ds_contrataciones/contrataciones_admin_2021/opendata/contratos.json"
@@ -55,7 +56,7 @@ def get_contract_id(contract):
 
 
 @print_progress
-def parse_contract(contract_id, language, contract):
+def parse_contract(contract_id, language, contract, update=False):
     try:
         os.makedirs(f"contracts/{contract_id}/{language}")
     except FileExistsError:
@@ -63,20 +64,27 @@ def parse_contract(contract_id, language, contract):
     data_xml_url = contract["dataXML"]
     metadata_xml_url = contract["metadataXML"]
 
-    with requests.get(data_xml_url) as r:
-        if r.ok:
-            with open(f"contracts/{contract_id}/{language}/data.xml", "wb") as f:
-                f.write(r.content)
+    if update or not os.path.exists(f"contracts/{contract_id}/{language}/data.xml"):
+        global REALLY_DOWNLOADED
+        REALLY_DOWNLOADED += 1
+        with requests.get(data_xml_url) as r:
+            if r.ok:
+                with open(f"contracts/{contract_id}/{language}/data.xml", "wb") as f:
+                    f.write(r.content)
 
-    with requests.get(metadata_xml_url) as r:
-        if r.ok:
-            with open(f"contracts/{contract_id}/{language}/metadata.xml", "wb") as f:
-                f.write(r.content)
+    if update or not os.path.exists(f"contracts/{contract_id}/{language}/metadata.xml"):
+        with requests.get(metadata_xml_url) as r:
+            if r.ok:
+                with open(
+                    f"contracts/{contract_id}/{language}/metadata.xml", "wb"
+                ) as f:
+                    f.write(r.content)
 
     contract["id"] = contract_id
 
-    with open(f"contracts/{contract_id}/{language}/data.json", "w") as f:
-        f.write(json.dumps(contract))
+    if update or not os.path.exists(f"contracts/{contract_id}/{language}/data.json"):
+        with open(f"contracts/{contract_id}/{language}/data.json", "w") as f:
+            f.write(json.dumps(contract))
 
 
 @print_progress
@@ -115,23 +123,35 @@ def merge_contracts(contracts_es, contracts_eu):
     return contracts
 
 
-def parse_multilingual_contract(contract_id, contract):
+def parse_multilingual_contract(contract_id, contract, update):
     for language, contract_data in contract.items():
-        parse_contract(contract_id, language, contract_data)
+        parse_contract(contract_id, language, contract_data, update)
 
 
-def get_contracts():
+def get_contracts(update):
     contracts_es = get_contracts_from_json(CONTRACTS_JSON_URL_ES)
     contracts_eu = get_contracts_from_json(CONTRACTS_JSON_URL_EU)
     contracts = merge_contracts(contracts_es, contracts_eu)
 
     count = 0
     for contract_id, contract in contracts.items():
-        parse_multilingual_contract(contract_id, contract)
+        parse_multilingual_contract(contract_id, contract, update)
         count += 1
         if count == LIMIT:
             break
 
 
 if __name__ == "__main__":
-    get_contracts()
+    parser = argparse.ArgumentParser(
+        description="Download contracts from the Euskadi Open Data portal"
+    )
+    parser.add_argument(
+        "--update",
+        action="store_true",
+        help="Update existing contracts",
+    )
+    myargs = parser.parse_args()
+
+    update = myargs.update
+
+    get_contracts(update)
